requirements.txt
pandas>=2.0.0
numpy>=1.24.0
python-dateutil>=2.8.2
openpyxl>=3.1.0
xlsxwriter>=3.1.0
PyQt6>=6.5.0
PyQt6-WebEngine>=6.5.0
dash>=2.14.0
dash-bootstrap-components>=1.5.0
plotly>=5.18.0
